+++ Source/Core/DolphinQt/GameList/GameListModel.cpp
@@ -172,7 +172,7 @@ QVariant GameListModel::data(const QModelIndex& index, int role) const
       return QString::fromStdString(str);
     }
     if (role == SORT_ROLE)
-      return static_cast<quint64>(game.GetFileSize());
+      return game.GetFileSize();
     break;
   case Column::FileFormat:
     if (role == Qt::DisplayRole || role == SORT_ROLE)
@@ -182,7 +182,7 @@ QVariant GameListModel::data(const QModelIndex& index, int role) const
     if (role == Qt::DisplayRole)
       return QString::fromStdString(UICommon::FormatSize(game.GetBlockSize()));
     if (role == SORT_ROLE)
-      return static_cast<quint64>(game.GetBlockSize());
+      return game.GetBlockSize();
     break;
   case Column::Compression:
     if (role == Qt::DisplayRole || role == SORT_ROLE)
@@ -207,7 +207,7 @@ QVariant GameListModel::data(const QModelIndex& index, int role) const
     if (role == SORT_ROLE)
     {
       const std::string game_id = game.GetGameID();
-      return static_cast<qlonglong>(m_timer.GetTimePlayed(game_id).count());
+      return m_timer.GetTimePlayed(game_id).count();
     }
     break;
   case Column::Tags:
